A bit of shader optimization and synchronization troubleshooting
I've spent 5 hours just learning about the command queue buffer in Unity and how to work with synchronization properly for the smoothingShader.

Most of tomorrow will likely be synchronization as well, as erosion is having some issues with that. The overall program may be more efficient if erosion comes before mesh creation, because mesh creation and normalization can happen simultaneously