Implemented range reduction shader because atomics don't work
I also figured out I never even had synchronization issues in the project. Unity's command buffer is really in-depth and almost completely prevents that as long as I'm not reading data on the cpu. The problem was the I had to manually set data for buffers on the cpu (Yay overhead) because values aren't reset to 0 on the gpu when you create a new buffer. This does happen in WebGPU, which is why those builds continued to work better than the editor.

Everythings is stable now in the editor again, which means it's time to look further into erosion optimizations.