Changed resolution and adjusted some parameters
Resolution changes were needed so the webGPU version could properly go fullscreen in itch.

A lot of this time was spent writing up instructions on itch so that laptop users can make sure their dedicated GPU is used rather than their integrated GPU, which has been the main bottleneck this entire time. This can only be handled on the user end because of security. It does let me increase the max values again though, as everyone's machines should be able to handle it