More erosion optimization
Drop distribution in the shader has fundamentally changed. That's probably because I spent several hours trying to create a tile system for better locality (Turns out even with that, constraining a certain number of drops per tile expands to clear tiled borders in erosion at large drop numbers). It's a bit more complex but I've gotten rid of the need for branching mostly in it. I believe this should also interface better with the GPU, I need to test that some more. I definitely need to test thread group size here, it's at the max size, which probably isn't a good thing.